[CREATE QUOTA - BACKEND] add base sceleton for section
Separated namespace, controllers, base views, routes so on.

[TRELLO CARD](https://trello.com/c/Kiospkws/361-dit-create-quota-iteration)
