+++ app/controllers/workbaskets/create_measures_controller.rb
@@ -1,76 +1,25 @@
 module Workbaskets
-  class CreateMeasuresController < Measures::BulksBaseController
+  class CreateMeasuresController < Workbaskets::BaseController
 
-    around_action :configure_time_machine
+    expose(:sub_klass) { "CreateMeasures" }
+    expose(:settings_type) { :create_measures }
 
-    before_action :require_to_be_workbasket_owner!,
-                  :require_step_declaration_in_params!,
-                  :check_if_action_is_permitted!,
-                  :status_check!, only: [ :edit, :update ]
-
-    before_action :handle_submit_for_cross_check!, only: [:update]
-
-    expose(:current_step) { params[:step] }
-
-    expose(:step_pointer) do
-      ::CreateMeasures::StepPointer.new(current_step)
-    end
-
-    expose(:previous_step) do
-      step_pointer.previous_step
-    end
-
-    expose(:workbasket_settings) do
-      workbasket.create_measures_settings
-    end
-
-    expose(:settings_params) do
-      ops = params[:settings]
-      ops.send("permitted=", true)
-      ops = ops.to_h
-
-      ops
-    end
-
-    expose(:saver_mode) { params[:mode] }
-
-    expose(:saver) do
-      Workbaskets::CreateMeasures::SettingsSaver.new(
-        workbasket,
-        current_step,
-        saver_mode,
-        settings_params
-      )
-    end
-
-    expose(:form) do
-      Workbaskets::CreateMeasures::Form.new(
-        Measure.new
-      )
-    end
-
-    expose(:attributes_parser) do
-      ::CreateMeasures::AttributesParser.new(
-        workbasket_settings,
-        current_step
-      )
-    end
-
-    expose(:submit_for_cross_check) do
-      ::Workbaskets::CreateMeasures::SubmitForCrossCheck.new(
-        workbasket
-      )
-    end
-
-    def new
-      self.workbasket = Workbaskets::Workbasket.buld_new_workbasket!(
-        :create_measures, current_user
-      )
-
-      redirect_to edit_create_measure_url(
+    expose(:initial_step_url) do
+      edit_create_measure_url(
         workbasket.id,
         step: :main
       )
+    end
+
+    expose(:previous_step_url) do
+      edit_create_measure_url(
+        workbasket.id,
+        step: previous_step
+      )
+    end
+
+    expose(:read_only_section_url) do
+      create_measure_url(workbasket.id)
     end
 
     def update
@@ -92,45 +41,11 @@ module Workbaskets
 
     private
 
-      def require_step_declaration_in_params!
-        if current_step.blank?
-          redirect_to edit_create_measure_url(
-            workbasket.id,
-            step: :main
-          )
-
-          return false
-        end
-      end
-
       def check_if_action_is_permitted!
         if step_pointer.review_and_submit_step? &&
            !workbasket_settings.validations_passed?(previous_step)
 
-          redirect_to edit_create_measure_url(
-            workbasket.id,
-            step: previous_step
-          )
-
-          return false
-        end
-      end
-
-      def status_check!
-        unless workbasket.in_progress?
-          redirect_to create_measure_url(workbasket.id)
-
-          return false
-        end
-      end
-
-      def handle_submit_for_cross_check!
-        if step_pointer.review_and_submit_step?
-          submit_for_cross_check.run!
-
-          render json: { redirect_url: create_measure_url(workbasket.id) },
-                 status: :ok
-
+          redirect_to previous_step_url
           return false
         end
       end
