+++ app/views/workbaskets/create_quota/_breadcrumbs.html.slim
@@ -1,0 +1,10 @@
+.breadcrumbs
+  ol
+    li
+      = link_to "Tariff Management", root_url
+
+    li
+      = link_to "Quotas", "#"
+
+    li aria-current="page"
+      | Create a quota
