+++ app/views/workbaskets/create_quota/edit.html.slim
@@ -1,0 +1,1 @@
+= render "workbaskets/shared/edit_base"
