+++ app/views/workbaskets/shared/_action_links.html.slim
@@ -21,5 +21,5 @@
   = link_to "Exit (return to main menu)", root_url, class: "secondary-button js-create-measures-exit-button"
 
   - if step_pointer.has_previous_step?
-    = link_to "Previous step", edit_create_measure_url(workbasket.id, step: previous_step),
+    = link_to "Previous step", previous_step_url,
                                class: "secondary-button js-create-measures-previous-step-link"
