+++ app/views/workbaskets/shared/_edit_base.html.slim
@@ -1,0 +1,18 @@
+= render "workbaskets/#{settings_type}/breadcrumbs"
+
+header
+  h1.heading-large
+    = public_send("#{settings_type}_section_header")
+
+- if step_pointer.current_step_is_form_step?
+  = render "workbaskets/#{settings_type}/form"
+- else
+  = render "workbaskets/#{settings_type}/steps/review_and_submit", read_only: false
+
+script
+  == "window.save_url = '/#{settings_type}/#{workbasket.id}';"
+  == "window.current_step = '#{current_step}';"
+  == "window.all_settings = #{workbasket_settings.settings.to_json};"
+
+  - unless step_pointer.review_and_submit_step?
+    == "window.current_step_settings = #{workbasket_settings.step_settings(current_step).to_json};"
