+++ app/views/workbaskets/shared/_js_variables.html.erb
