+++ app/views/workbaskets/shared/_success_message.html.slim
